add parameter to delay transmission
Using '-T' we add the specified delay (in ns) after each character. This
helps when you have no flow control and lose characters because of
overflowing buffers.

This has been requested with issue #96 and ithe first version was sent
with PR #131 by Heinrich Schuchardt. This code is based on it.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>
Signed-off-by: Wolfram Sang <wsa@kernel.org>