build: Use simple (non-annotated) tags for development version
When building from git, we use `git describe --long` to generate a
version string. This only considers annotated tags, so everything gets
a version relative to v3.2.0.

This commit adds `--tags` to the command line so that simple tags
(e.g., those added by semantic-release) are considered as well.

This means that instead of:

    v3.2.0-65-gb188523

We get:

    v4.0.1-4-gb188523